requirements.txt
langchain>=0.0.20
langchain-anthropic>=0.1.0
langchain-community >=0.1.0
langchain-openai>=0.1.0
pydantic>=2.0.3
wikipedia
duckduckgo-search
langchain-core>=0.1.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.0.20
anthropic>=0.8.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
redis>=5.0.0
pymongo>=4.5.0
motor>=3.3.0
numpy>=1.24.0
networkx>=3.0.0
rdflib>=6.0.0
scikit-learn>=1.2.0
ollama>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Document Processing Pipeline Dependencies
httpx>=0.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0
html2text>=2020.1.16
python-magic>=0.4.27
chardet>=5.0.0
langdetect>=1.0.9
textstat>=0.7.3

# Excel and PowerPoint support
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-pptx>=0.6.21

# Vectorization and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0

# Vector stores
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Web interface
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0

# Progress bars
tqdm>=4.65.0

# Memory optimization
psutil>=5.9.0

# API Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-multipart>=0.0.6
starlette>=0.27.0
jinja2>=3.1.2
websockets>=11.0.3

# Data Processing Dependencies
pandas>=2.0.0
polars>=0.20.0
dask>=2023.12.0
apache-airflow>=2.8.0
kafka-python>=2.0.2
confluent-kafka>=2.3.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
influxdb-client>=1.40.0
minio>=7.2.0
great-expectations>=0.18.0
structlog>=23.2.0
prometheus-client>=0.19.0
celery>=5.3.0
schedule>=1.2.0
croniter>=2.0.0
pyarrow>=14.0.0
fsspec>=2023.12.0

# Enhanced DataMCPServerAgent Dependencies
# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0

# Email Integration
aiosmtplib>=3.0.1
sendgrid>=6.10.0
mailgun>=0.1.1

# WebRTC & Communication
aiortc>=1.6.0

# Cloudflare Integration
cloudflare>=2.11.1

# Database Extensions
alembic>=1.13.1
aiosqlite>=0.19.0

# Monitoring & Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0

# Configuration & Utilities
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
typer>=0.9.0
python-dateutil>=2.8.2

# Image & Audio Processing
pillow>=10.1.0
pydub>=0.25.1

# Security & Encryption
cryptography>=41.0.8

# Rate Limiting & Caching
slowapi>=0.1.9
aiocache>=0.12.2

# System Monitoring
psutil>=5.9.6

# JSON & Data Handling
orjson>=3.9.10
shortuuid>=1.0.11
marshmallow>=3.20.1
marshmallow-dataclass>=8.6.0

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
factory-boy>=3.3.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.6.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.8

# Deployment
docker>=6.1.3
kubernetes>=28.1.0

# Async Utilities
asyncio-mqtt>=0.16.1
aio-pika>=9.3.1

# Dependency Injection & Patterns
dependency-injector>=4.41.0
transitions>=0.9.0
tenacity>=8.2.3

# Time Zone Handling
pytz>=2023.3

# Development Tools
watchfiles>=0.21.0
supervisor>=4.2.5
memory-profiler>=0.61.0
py-spy>=0.3.14
locust>=2.17.0

# Document Processing Dependencies
pypdf2>=3.0.0
python-docx>=1.1.0
python-magic>=0.4.27
chardet>=5.2.0
markdown>=3.5.0
html2text>=2020.1.16
python-pptx>=0.6.23
openpyxl>=3.1.0
xlrd>=2.0.1

# Text Processing and NLP
spacy>=3.7.0
nltk>=3.8.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
transformers>=4.35.0
textstat>=0.7.3
langdetect>=1.0.9
unidecode>=1.3.7

# Vector Stores and Embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4
pinecone-client>=2.2.0
weaviate-client>=3.25.0
openai>=1.3.0
qdrant-client>=1.6.0
milvus>=2.3.0

# Text Chunking and Processing
semantic-text-splitter>=0.8.0
langchain-text-splitters>=0.0.1

# Penetration Testing Dependencies
python-nmap>=0.7.1
shodan>=1.28.0
censys>=2.2.0
dnspython>=2.4.0
python-whois>=0.8.0
ipaddress>=1.0.23
validators>=0.22.0
netaddr>=0.9.0

# Security and Sandboxing
docker>=6.1.3
firejail>=0.9.72
seccomp>=0.1.0

# Additional OSINT Tools
requests-html>=0.10.0
selenium>=4.15.0
scrapy>=2.11.0
theHarvester>=4.4.0

# Vulnerability Databases
python-cve>=1.0.0
vulners>=2.0.0

# Network Analysis
scapy>=2.5.0
impacket>=0.11.0

# Reporting and Visualization
jinja2>=3.1.2
matplotlib>=3.7.0
plotly>=5.17.0
reportlab>=4.0.0

# Algorithmic Trading Dependencies
# Machine Learning for Trading Strategies
tensorflow>=2.13.0
scipy>=1.11.0
joblib>=1.3.0

# Financial and Technical Analysis
ta-lib>=0.4.25
yfinance>=0.2.18
ccxt>=4.0.0
quantlib>=1.31
zipline-reloaded>=3.0.0

# Trading Visualization
seaborn>=0.12.0
mplfinance>=0.12.0

# WebSocket for Real-time Data
websockets>=11.0.0
aiohttp>=3.8.0

# Additional Trading Utilities
python-binance>=1.0.19
alpha-vantage>=2.3.1
polygon-api-client>=1.12.0

# Enhanced Reinforcement Learning Dependencies
gymnasium>=0.28.0
stable-baselines3>=2.0.0
tensorboard>=2.8.0
wandb>=0.15.0
optuna>=3.0.0
ray[rllib]>=2.5.0
torch-geometric>=2.3.0
torchvision>=0.15.0
torchtext>=0.15.0
higher>=0.2.1

# Cloud Integration
boto3>=1.34.0
azure-identity>=1.15.0
azure-mgmt-compute>=30.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0

# Federated Learning & Privacy
pycryptodome>=3.19.0
cryptography>=41.0.8

# Auto-Scaling & Monitoring
websockets>=11.0.3
aiohttp>=3.9.0
